refactor: EngineManifest 系モジュールを統合
